Add loop process that wait for the DB is ready to accept connections.
Before, there was only "sleep 3" command for waiting
for the DB is ready to accept connections. However,
sometimes, DB is not ready to accept to connections
within 3 second. In that case some errors occur in
the after processes. To prevent this issue, add loop
process that wait for the DB is ready to accept by
using "SELECT 1".